Make xkas.dll bundle (correctly) into the nupkg file
This was a wild ride.
Ultimately I needed to use the .targets file to set the build properties of the xkas.dll so it copied appropriately to output.
So it would be available next to the .dll files that consume it.
But it could NOT be added like the other DLLs, because reasons relating to it being an unmanaged DLL(?).